Added 2 functions per button
The timing for the pump and heating element is still in testing.
Added library for the functionality of the buttons. Now when press short coffee it will run the pump so it fill when  its dry, and the large coffee will act as a pre-heat for the water.
The code was further organized and optimized